Refactor in response to comments
Instead of deferring feature computation, compute initial features at
option parse time. Provide a separate computeFeatures method that
takes a module and updates the features to respect the target features
section.

Also move readFeaturesSection to module-utils.h, which lets me put the
LEB support code back where it was in wasm-binary.h.
