+++ src/support/command-line.h
@@ -34,7 +34,7 @@ namespace wasm {
 
 class Options {
  public:
-  typedef std::function<void(Options *, const std::string& )> Action;
+  using Action = std::function<void(Options *, const std::string& )>;
   enum class Arguments { Zero, One, N, Optional };
 
   bool debug;
@@ -44,7 +44,7 @@ class Options {
   ~Options();
   Options &add(const std::string& longName, const std::string& shortName,
                const std::string& description, Arguments arguments,
-               const Action &action);
+               const Action& action);
   Options &add_positional(const std::string& name, Arguments arguments,
                           const Action &action);
   void parse(int argc, const char *argv[]);
