+++ src/tools/optimization-options.h
@@ -130,7 +130,7 @@ struct OptimizationOptions : public ToolOptions {
   void runPasses(Module& wasm) {
     PassRunner passRunner(&wasm, passOptions);
     if (debug) passRunner.setDebug(true);
-    passRunner.setFeatures(getFeatures(wasm));
+    passRunner.setFeatures(passOptions.features);
     for (auto& pass : passes) {
       if (pass == DEFAULT_OPT_PASSES) {
         passRunner.addDefaultOptimizationPasses();
