+++ src/tools/tool-options.h
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+#include "ir/module-utils.h"
 #include "support/command-line.h"
 #include "pass.h"
 
@@ -24,31 +25,41 @@
 namespace wasm {
 
 struct ToolOptions : public Options {
-  using Builder = std::function<void (FeatureSet&)>;
   PassOptions passOptions;
 
   ToolOptions(const std::string& command, const std::string& description)
       : Options(command, description) {
     (*this)
-        .addFeatureOpt("--mvp-features", "-mvp", "Disable all non-MVP features",
-                       [](FeatureSet& features) {
-                         features = FeatureSet::MVP;
-                       })
-        .addFeatureOpt("--all-features", "-all",
-                       "Enable all features "
-                       "(default if there is no target features section"
-                       " or if any feature options are provided)",
-                       [](FeatureSet& features) {
-                         features = FeatureSet::All;
-                       })
-        .addFeatureOpt("--detect-features", "",
-                       "Use features from the target features section"
-                       "(default if there is a target features section"
-                       " and no feature options are provided)",
-                       [this](FeatureSet& features) {
-                         features = wasmFeatures;
-                         errorOnFeatureMismatch = false;
-                       })
+        .add("--mvp-features", "-mvp", "Disable all non-MVP features",
+             Arguments::Zero,
+             [this](Options*, const std::string&) {
+               hasFeatureOptions = true;
+               passOptions.features.makeMVP();
+               enabledFeatures.makeMVP();
+               disabledFeatures.setAll();
+             })
+        .add("--all-features", "-all", "Enable all features "
+             "(default if there is no target features section"
+             " or if any feature options are provided)",
+             Arguments::Zero,
+             [this](Options*, const std::string&) {
+               hasFeatureOptions = true;
+               passOptions.features.setAll();
+               enabledFeatures.setAll();
+               disabledFeatures.makeMVP();
+             })
+        .add("--detect-features", "",
+             "Use features from the target features section"
+             "(default if there is a target features section"
+             " and no feature options are provided)",
+             Arguments::Zero,
+             [this](Options*, const std::string&) {
+               hasFeatureOptions = true;
+               detectFeatures = true;
+               enabledFeatures.makeMVP();
+               disabledFeatures.makeMVP();
+             });
+    (*this)
         .addFeature("sign-ext", "sign extension operations",
                     FeatureSet::SignExt)
         .addFeature("threads", "atomic operations",
@@ -70,63 +81,54 @@ struct ToolOptions : public Options {
              });
   }
 
-  ToolOptions &addFeatureOpt(const std::string& longName,
-                             const std::string& shortName,
-                             const std::string& description,
-                             Builder builder) {
-    add(longName, shortName, description, Options::Arguments::Zero,
-        [=](Options*, const std::string&) {
-          featureBuilders.push_back(builder);
-        });
-    return *this;
-  }
-
-  ToolOptions &addFeature(const std::string& name,
+  ToolOptions& addFeature(const std::string& name,
                           const std::string& description,
                           FeatureSet::Feature feature) {
     (*this)
-        .addFeatureOpt(std::string("--enable-") + name, "",
-                       std::string("Enable ") + description,
-                       [=](FeatureSet& features) {
-                         features.set(feature, true);
-                       })
-        .addFeatureOpt(std::string("--disable-") + name, "",
-                       std::string("Disable ") + description,
-                       [=](FeatureSet& features) {
-                         features.set(feature, false);
-                       });
+        .add(std::string("--enable-") + name, "",
+             std::string("Enable ") + description, Arguments::Zero,
+             [=](Options*, const std::string&) {
+               hasFeatureOptions = true;
+               passOptions.features.set(feature, true);
+               enabledFeatures.set(feature, true);
+               disabledFeatures.set(feature, false);
+             })
+
+        .add(std::string("--disable-") + name, "",
+             std::string("Disable ") + description, Arguments::Zero,
+             [=](Options*, const std::string&) {
+               hasFeatureOptions = true;
+               passOptions.features.set(feature, false);
+               enabledFeatures.set(feature, false);
+               disabledFeatures.set(feature, true);
+             });
     return *this;
   }
 
-  FeatureSet& getFeatures(const Module& module) {
-    if (featuresInitialized) {
-      return passOptions.features;
-    }
+  void calculateFeatures(const Module& module) {
+    FeatureSet wasmFeatures;
+    bool wasmHasFeatures =
+        ModuleUtils::readFeaturesSection(module, wasmFeatures);
 
-    bool wasmHasFeatures = module.readFeatures(wasmFeatures);
-    if (featureBuilders.size()) {
-      for (auto& builder : featureBuilders) {
-        builder(passOptions.features);
-      }
-      if (errorOnFeatureMismatch &&
-          !(wasmFeatures <= passOptions.features)) {
+    if (hasFeatureOptions) {
+      if (detectFeatures) {
+        wasmFeatures.enable(enabledFeatures);
+        wasmFeatures.disable(disabledFeatures);
+        passOptions.features = wasmFeatures;
+      } else if (!(wasmFeatures <= passOptions.features)) {
         Fatal() << "module uses features not explicitly specified, "
                 << "use --detect-features to resolve";
       }
     } else if (wasmHasFeatures) {
       passOptions.features = wasmFeatures;
     }
-    featuresInitialized = true;
-    return passOptions.features;
   }
 
 private:
-  std::vector<Builder> featureBuilders;
-  bool errorOnFeatureMismatch = true;
-  bool featuresInitialized = false;
-
-  // cached contents of the target features section
-  FeatureSet wasmFeatures = FeatureSet::MVP;
+  bool hasFeatureOptions = false;
+  bool detectFeatures = false;
+  FeatureSet enabledFeatures = FeatureSet::MVP;
+  FeatureSet disabledFeatures = FeatureSet::MVP;
 };
 
 } // namespace wasm
