+++ src/tools/wasm2js.cpp
@@ -81,6 +81,7 @@ int main(int argc, const char *argv[]) {
       ModuleReader reader;
       reader.setDebug(options.debug);
       reader.read(input, wasm, "");
+      options.calculateFeatures(wasm);
     } else {
       auto input(
           read_file<std::vector<char>>(options.extra["infile"], Flags::Text, options.debug ? Flags::Debug : Flags::Release));
@@ -99,7 +100,7 @@ int main(int argc, const char *argv[]) {
   }
 
   if (options.passOptions.validate) {
-    if (!WasmValidator().validate(wasm, options.getFeatures(wasm))) {
+    if (!WasmValidator().validate(wasm, options.passOptions.features)) {
       WasmPrinter::printModule(&wasm);
       Fatal() << "error in validating input";
     }
