+++ src/wasm.h
@@ -954,9 +954,6 @@ public:
 
   void updateMaps();
 
-  // returns true if the module has a target features section
-  bool readFeatures(FeatureSet&) const;
-
   void clearDebugInfo();
 };
