+++ src/wasm/wasm.cpp
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-#include "leb.h"
 #include "wasm.h"
 #include "wasm-traversal.h"
 #include "ir/branch-utils.h"
@@ -971,72 +970,6 @@ void Module::updateMaps() {
   for (auto& curr : globals) {
     globalsMap[curr->name] = curr.get();
   }
-}
-
-bool Module::readFeatures(FeatureSet& features) const try {
-  features = FeatureSet::MVP;
-
-  const UserSection* section = nullptr;
-  for (auto &s : userSections) {
-    if (s.name == BinaryConsts::UserSections::TargetFeatures) {
-      section = &s;
-      break;
-    }
-  }
-
-  if (!section) {
-    return false;
-  }
-
-  // read target features section
-  size_t index = 0;
-  auto next_byte = [&]() {
-    return section->data.at(index++);
-  };
-
-  size_t num_feats = U32LEB().read(next_byte).value;
-  for (size_t i = 0; i < num_feats; ++i) {
-    uint8_t prefix = section->data.at(index++);
-    if (prefix != '=' && prefix != '+' && prefix != '-') {
-      // unrecognized prefix, silently fail
-      features = FeatureSet::MVP;
-      return false;
-    }
-
-    size_t len = U32LEB().read(next_byte).value;
-    if (index + len > section->data.size()) {
-      // ill-formed string, silently fail
-      features = FeatureSet::MVP;
-      return false;
-    }
-    std::string name(&section->data[index], len);
-    index += len;
-
-    if (prefix == '-') {
-      continue;
-    }
-
-    if (name == BinaryConsts::UserSections::AtomicsFeature) {
-      features.setAtomics();
-    } else if (name == BinaryConsts::UserSections::BulkMemoryFeature) {
-      features.setBulkMemory();
-    } else if (name == BinaryConsts::UserSections::ExceptionHandlingFeature) {
-      WASM_UNREACHABLE(); // TODO: exception handling
-    } else if (name == BinaryConsts::UserSections::TruncSatFeature) {
-      features.setTruncSat();
-    } else if (name == BinaryConsts::UserSections::SignExtFeature) {
-      features.setSignExt();
-    } else if (name == BinaryConsts::UserSections::SIMD128Feature) {
-      features.setSIMD();
-    }
-  }
-
-  return true;
-
-// silently fail to read features. Maybe this should warn?
-} catch (std::out_of_range& e) {
-  features = FeatureSet::MVP;
-  return false;
 }
 
 void Module::clearDebugInfo() {
